# docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "汽车健康管家平台 S0 API 草案",
    "version": "0.1.0",
    "description": "43 个来源核心接口加 S0 支撑接口。未明确的业务字段以 S0_BUSINESS_CONTRACT.md 为准；未实现的路径不代表已可调用。"
  },
  "servers": [
    {
      "url": "http://127.0.0.1:8080"
    }
  ],
  "paths": {
...
    "/api/order/create": {
      "post": {
        "summary": "创建本人待支付预约",
        "operationId": "post_api_order_create",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/CreatedReservation"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "merchant_project_id",
                  "quote_version_id",
                  "vehicle_id",
                  "slot_id"
                ],
                "properties": {
                  "merchant_project_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "quote_version_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "vehicle_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "slot_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  }
                }
              }
            }
          }
        },
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
    "/api/order/list": {
      "get": {
        "summary": "本人订单分页",
        "operationId": "get_api_order_list",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "status",
            "in": "query",
            "schema": {
              "type": "string",
              "enum": [
                "PENDING_PAYMENT",
                "PAID",
                "CLOSED"
              ]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OwnerOrderPage"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
    "/api/order/{id}": {
      "get": {
        "summary": "本人订单及下单快照",
        "operationId": "get_api_order_id",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OwnerOrderDetail"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
    "/api/order/cancel": {
      "post": {
        "summary": "取消本人待支付订单",
        "operationId": "post_api_order_cancel",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/OwnerOrderDetail"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "order_id"
                ],
                "properties": {
                  "order_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  }
                }
              }
            }
          }
        },
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
...
    "/api/order/slots": {
      "get": {
        "summary": "可预约项目时段分页",
        "operationId": "get_api_order_slots",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "merchant_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "project_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "date",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationSlotPage"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
...
    "/api/merchant/slots": {
      "get": {
        "summary": "本店预约时段分页",
        "operationId": "get_api_merchant_slots",
        "tags": [
          "merchant"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "page",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 1000000,
              "default": 1
            }
          },
          {
            "name": "page_size",
            "in": "query",
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 100,
              "default": 20
            }
          },
          {
            "name": "date",
            "in": "query",
            "schema": {
              "type": "string",
              "format": "date"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationSlotPage"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "MERCHANT",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      },
      "post": {
        "summary": "发布本店项目时段",
        "operationId": "post_api_merchant_slots",
        "tags": [
          "merchant"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationSlot"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [
                  "standard_project_id",
                  "starts_at",
                  "ends_at",
                  "capacity"
                ],
                "properties": {
                  "standard_project_id": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 9007199254740991
                  },
                  "starts_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "ends_at": {
                    "type": "string",
                    "format": "date-time"
                  },
                  "capacity": {
                    "type": "integer",
                    "minimum": 1,
                    "maximum": 100
                  }
                }
              }
            }
          }
        },
        "x-roles": "MERCHANT",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
    "/api/merchant/slots/{id}/close": {
      "post": {
        "summary": "关闭本店预约时段",
        "operationId": "post_api_merchant_slots_id_close",
        "tags": [
          "merchant"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          },
          {
            "name": "Idempotency-Key",
            "in": "header",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationSlot"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "additionalProperties": false,
                "required": [],
                "properties": {}
              }
            }
          }
        },
        "x-roles": "MERCHANT",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
    "/api/order/quote/{id}": {
      "get": {
        "summary": "当前预约报价确认上下文",
        "operationId": "get_api_order_quote_id",
        "tags": [
          "order"
        ],
        "x-feature-ids": [
          "F07",
          "F08"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "integer",
              "minimum": 1,
              "maximum": 9007199254740991
            }
          }
        ],
        "responses": {
          "200": {
            "description": "统一成功响应",
            "content": {
              "application/json": {
                "schema": {
                  "allOf": [
                    {
                      "$ref": "#/components/schemas/ApiResponse"
                    },
                    {
                      "type": "object",
                      "properties": {
                        "data": {
                          "$ref": "#/components/schemas/ReservationQuote"
                        }
                      }
                    }
                  ]
                }
              }
            }
          },
          "400": {
            "description": "参数或业务状态错误",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiResponse"
                }
              }
            }
          },
          "401": {
            "description": "认证失败，code=40100"
          },
          "403": {
            "description": "无权访问，code=40300"
          },
          "404": {
            "description": "本人/本店资源不可用，code=40400"
          },
          "409": {
            "description": "40901报价版本改变、40902时段关闭/开始、40903名额不足、40904时段重叠"
          },
          "503": {
            "description": "数据库或事务暂不可用，code=50300"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "x-roles": "OWNER",
        "x-implementation-status": "local-reservations-core-implemented",
        "description": "本机预约订单子集，支付基础接入、正式扣款/券/退款未接入；时间带时区、数据库UTC，日历按北京时间。详见 RESERVATION_ORDERS.md。"
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    },
    "schemas": {
      "ApiResponse": {
        "type": "object",
        "required": [
          "code",
          "message",
          "request_id"
        ],
        "properties": {
          "code": {
            "type": "integer"
          },
          "message": {
            "type": "string"
          },
          "data": {},
          "request_id": {
            "type": "string",
            "format": "uuid"
          }
        }
      },
...
      "ReservationSlotPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ReservationSlot"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        }
      },
      "OwnerOrderPage": {
        "type": "object",
        "required": [
          "items",
          "total",
          "page",
          "page_size"
        ],
        "properties": {
          "items": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/OwnerOrder"
            }
          },
          "total": {
            "type": "integer",
            "minimum": 0
          },
          "page": {
            "type": "integer",
            "minimum": 1
          },
          "page_size": {
            "type": "integer",
            "minimum": 1,
            "maximum": 100
          }
        }
      },
...
      "OwnerOrderDetail": {
        "type": "object",
        "required": [
          "order_id",
          "order_no",
          "amount_due",
          "status",
          "expires_at",
          "created_at",
          "closed_at",
          "close_reason",
          "project_snapshot",
          "merchant_snapshot",
          "appointment_snapshot",
          "vehicle_id",
          "price_snapshot",
          "payment_summary"
        ],
        "properties": {
          "order_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "order_no": {
            "type": "string"
          },
          "amount_due": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,7})\\.[0-9]{2}$"
          },
          "status": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "created_at": {
            "type": "string",
            "format": "date-time"
          },
          "closed_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          },
          "close_reason": {
            "type": [
              "string",
              "null"
            ]
          },
          "project_snapshot": {
            "type": [
              "object",
              "null"
            ]
          },
          "merchant_snapshot": {
            "type": [
              "object",
              "null"
            ]
          },
          "appointment_snapshot": {
            "type": [
              "object",
              "null"
            ]
          },
          "vehicle_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "price_snapshot": {
            "type": [
              "object",
              "null"
            ]
          },
          "payment_summary": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/PaymentSummary"
              },
              {
                "type": "null"
              }
            ]
          }
        }
      },
      "CreatedReservation": {
        "type": "object",
        "required": [
          "order_id",
          "order_no",
          "amount_due",
          "status",
          "expires_at"
        ],
        "properties": {
          "order_id": {
            "type": "integer",
            "minimum": 1,
            "maximum": 9007199254740991
          },
          "order_no": {
            "type": "string"
          },
          "amount_due": {
            "type": "string",
            "pattern": "^(?:0|[1-9][0-9]{0,7})\\.[0-9]{2}$"
          },
          "status": {
            "type": "string"
          },
          "expires_at": {
            "type": [
              "string",
              "null"
            ],
            "format": "date-time"
          }
        }
      },
...
    }
  }
}